requirements.txt
alembic==1.7.7
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==1.1.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.0
markdown2==2.3.8
MarkupSafe==2.0.1
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==3.0.1
zipp==3.8.0
